Prefer vendored esptool over a system-installed esptool when flashing
On Linux the binary flasher (flasher.pyz / binary_configurator.py) runs
under the system Python, so an incompatible esptool in site-packages
(e.g. Fedora 44's esptool 5.x, which removed esptool.cmds.make_image)
shadowed the esptool vendored by the firmware. Point PYTHONPATH at the
firmware checkout's python/external/esptool for the flasher spawn so the
matching vendored copy always wins.

Also apply the PYTHONPATH/PYTHONHOME env blacklist (#440) in the Python
class, which previously received the unfiltered environment.

Fixes #780
