+++ src/api/src/library/Platformio/index.ts
@@ -7,6 +7,7 @@ import Commander, { CommandResult, NoOpFunc, OnOutputFunc } from '../Commander';
 import { LoggerService } from '../../logger';
 import UploadType from './Enum/UploadType';
 import Python from '../Python';
+import envFilter from '../envFilter';
 
 interface PlatformioCoreState {
   core_version: string;
@@ -30,19 +31,6 @@ const prependPATH = (pth: string, item: string): string => {
     return `${item}${path.delimiter}${pth}`;
   }
   return item;
-};
-
-const envFilter = (
-  env: NodeJS.ProcessEnv,
-  blacklist: string[],
-): NodeJS.ProcessEnv => {
-  const result: NodeJS.ProcessEnv = {};
-  Object.keys(env).forEach((key) => {
-    if (!blacklist.includes(key)) {
-      result[key] = env[key];
-    }
-  });
-  return result;
 };
 
 export default class Platformio {
